+++ src/main/java/dev/nyon/bbm/asm/AbstractBoatMixin.java
@@ -2,7 +2,6 @@ package dev.nyon.bbm.asm;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import dev.nyon.bbm.BbmBoat;
-import dev.nyon.bbm.KeyBindings;
 import dev.nyon.bbm.config.Config;
 import dev.nyon.bbm.config.ConfigKt;
 import net.minecraft.core.BlockPos;
@@ -10,7 +9,7 @@ import net.minecraft.tags.BlockTags;
 import net.minecraft.util.Mth;
 import net.minecraft.world.entity.player.Player;
 //? if >=1.21.3
-import net.minecraft.world.entity.vehicle.AbstractBoat;
+/*import net.minecraft.world.entity.vehicle.AbstractBoat;*/
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
@@ -20,9 +19,7 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Constant;
-import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyConstant;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -46,7 +43,7 @@ public class AbstractBoatMixin implements BbmBoat {
     }
 
     /*? if >=1.21.3 {*/
-    @Shadow
+    /*@Shadow
     private AbstractBoat.Status status;
     @Unique
     private AbstractBoat instance = (AbstractBoat) (Object) this;
@@ -143,23 +140,6 @@ public class AbstractBoatMixin implements BbmBoat {
         return config.getPlayerEjectTicks();
     }
 
-    @Inject(
-        method = "tick",
-        at = @At("HEAD")
-    )
-    private void triggerJump(CallbackInfo ci) {
-        Config config = ConfigKt.getActiveConfig();
-        if (config == null) return;
-        if (!KeyBindings.INSTANCE.getJumpKeyBind().isDown() || !config.getAllowJumpKeybind()) return;
-        if (
-            config.getOnlyKeybindJumpOnGroundOrWater()
-                && !instance.onGround()
-                && !instance.isInWater()
-                && !instance.isUnderWater()
-        ) return;
-        instance.addDeltaMovement(new Vec3(0.0, config.getStepHeight() * config.getKeybindJumpHeightMultiplier(), 0.0));
-    }
-
     @Unique
     private boolean failsPlayerCondition() {
         Config config = ConfigKt.getActiveConfig();
@@ -169,5 +149,5 @@ public class AbstractBoatMixin implements BbmBoat {
         return instance.getPassengers().stream()
             .noneMatch(entity -> entity instanceof Player);
     }
-    /*?}*/
+    *//*?}*/
 }