+++ src/main/java/dev/nyon/bbm/asm/BoatMixin.java
@@ -2,7 +2,6 @@ package dev.nyon.bbm.asm;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import dev.nyon.bbm.BbmBoat;
-import dev.nyon.bbm.KeyBindings;
 import dev.nyon.bbm.config.Config;
 import dev.nyon.bbm.config.ConfigKt;
 import net.minecraft.core.BlockPos;
@@ -21,9 +20,7 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Constant;
-import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyConstant;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -52,7 +49,7 @@ abstract class BoatMixin extends Entity implements BbmBoat {
     }
 
     /*? if <1.21.3 {*/
-    /*@Shadow
+    @Shadow
     private Boat.Status status;
 
     @Unique
@@ -147,23 +144,6 @@ abstract class BoatMixin extends Entity implements BbmBoat {
         return config.getPlayerEjectTicks();
     }
 
-    @Inject(
-        method = "tick",
-        at = @At("HEAD")
-    )
-    private void triggerJump(CallbackInfo ci) {
-        Config config = ConfigKt.getActiveConfig();
-        if (config == null) return;
-        if (!KeyBindings.INSTANCE.getJumpKeyBind().isDown() || !config.getAllowJumpKeybind()) return;
-        if (
-            config.getOnlyKeybindJumpOnGroundOrWater()
-                && !onGround()
-                && !isInWater()
-                && !isUnderWater()
-        ) return;
-        addDeltaMovement(new Vec3(0.0, config.getStepHeight() * config.getKeybindJumpHeightMultiplier(), 0.0));
-    }
-
     @Unique
     private boolean failsPlayerCondition() {
         Config config = ConfigKt.getActiveConfig();
@@ -173,5 +153,5 @@ abstract class BoatMixin extends Entity implements BbmBoat {
         return getPassengers().stream()
             .noneMatch(entity -> entity instanceof Player);
     }
-    *//*?}*/
+    /*?}*/
 }