+++ src/main/kotlin/dev/nyon/bbm/KeyBindings.kt
@@ -10,4 +10,14 @@ object KeyBindings {
             "key.bbm.jump", InputConstants.Type.KEYSYM, GLFW.GLFW_KEY_H, "key.bbm.category"
         )
     }
+
+    fun register() {
+        //? if fabric {
+        net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper.registerKeyBinding(jumpKeyBind)
+        //?} else {
+        /*dev.nyon.klf.MOD_BUS.addListener<net.neoforged.neoforge.client.event.RegisterKeyMappingsEvent> {
+            it.register(jumpKeyBind)
+        }
+        *///?}
+    }
 }